<test> 구간 테스트 파일 위치 변경, 리팩토링
